Fix the background loop for continuous queries to execute every second
While it is a common pattern to use `<-time.After(time.Second)` in a
select loop, it has a nasty side effect that it creates a channel that
continues to exist even if you don't wait the full time.

It also means that if a bunch of queries are requested by the user, it
will reset the timer for the normal per second run.

This change moves the channel to an outer variable so the next for loop
will use the same channel and will execute after a second instead of
resetting every time a run is requested.